==> project/p3-Evaluation.ipynb <==
import os
import sys
sys.path.append('../')
import pickle
import pandas as pd
import numpy as np

# Evaluation of CF Recommender Systems

# TEST SET
df_test = pd.read_pickle("files/testset.pkl")[["reviewerID", "asin", "overall"]]\
            .rename(columns={"reviewerID": "uid", "asin": "iid"})

# PREDICTIONS
nb = pd.read_pickle("files/preds_knn.pkl")
lf = pd.read_pickle("files/preds_svd.pkl")
pred_nb_list = list(nb.itertuples(index=False))
pred_lf_list = list(lf.itertuples(index=False))

# Detect users from training set that are not in test
nb_users = set([pred.uid for pred in pred_nb_list])
lf_users = set([pred.uid for pred in pred_lf_list])
nb_users_in_pred_but_not_in_test = list(nb_users.difference(set(df_test['uid'])))
lf_users_in_pred_but_not_in_test = list(lf_users.difference(set(df_test['uid'])))
assert nb_users_in_pred_but_not_in_test == lf_users_in_pred_but_not_in_test
print(f"There are {len(lf_users_in_pred_but_not_in_test)} users in the training set that are not in the test set.")

# Remove these users' predictions for evaluation
print("Lengths before removing preds not in test set:", len(nb), len(lf))
nb = nb[~nb.uid.isin(nb_users_in_pred_but_not_in_test)]
lf = lf[~lf.uid.isin(nb_users_in_pred_but_not_in_test)]
print("After removing:", len(nb), len(lf))

nb_merge = nb.merge(df_test, how="inner", on=["uid", "iid"])
print("\nkNN RMSE:", np.sqrt(np.mean((nb_merge["overall"] - nb_merge["score"])**2)))

lf_merge = lf.merge(df_test, how="inner", on=["uid", "iid"])
print("SVD RMSE:", np.sqrt(np.mean((lf_merge["overall"] - lf_merge["score"])**2)))

# often used functions (see metrics.py; rather concise implementations)
from metrics import PatK, MAPatK, MRRatK, HRatK

ks = [5, 15]

print(12*" " + "NB | LF")

for k in ks:
    P_nb, P_lf = PatK(nb, df_test, k), PatK(lf, df_test, k)
    MAP_nb, MAP_lf = MAPatK(nb, df_test, k), MAPatK(lf, df_test, k)
    MRR_nb, MRR_lf = MRRatK(nb, df_test, k), MRRatK(lf, df_test, k)
    MHR_nb, MHR_lf = HRatK(nb, df_test, k), HRatK(lf, df_test, k)
    print(f"  P@{k:2g} = {P_nb  :.4f}|{P_lf  :.4f}")
    print(f"MAP@{k:2g} = {MAP_nb:.4f}|{MAP_lf:.4f}")
    print(f"MRR@{k:2g} = {MRR_nb:.4f}|{MRR_lf:.4f}")
    print(f" HR@{k:2g} = {MHR_nb:.4f}|{MHR_lf:.4f}\n") 


## Error analysis

from metrics import relevant_column, first

# identify users where model performs badly
train = pd.read_pickle("files/train.pkl")[["reviewerID","asin","overall"]]
merged = relevant_column(nb, df_test, 30)
scores = merged[["uid", "iid", "relevant"]].groupby(by="uid").apply(first)

# calculate statistics for these users 
ids = scores[(scores != 0)].index
train.groupby("reviewerID").size()[ids].mean(), train.groupby("reviewerID").size().mean()

good_asins = train[train["reviewerID"].isin(ids)].asin.unique()
train.groupby("asin").size()[good_asins].mean(), train.groupby("asin").size().mean()

# identify users ('first' part of project description error analysis)
dft = pd.read_pickle("files/testset.pkl")[["reviewerID", "asin", "overall", "unixReviewTime"]] \
        .sort_values("unixReviewTime") \
        .reset_index(drop=True)

uid_first, uid_last = dft.iloc[[0,-1]]["reviewerID"]
uid_first, uid_last


df_train = pd.read_pickle("files/train.pkl")
df = df_train[["overall", "reviewerID", "asin"]]

df_with_nans = df.pivot_table(values="overall", index="reviewerID", columns="asin")
df = df_with_nans.fillna(0)

corrs = df.T.corr("pearson")

first_nbs = corrs.loc[uid_first].nlargest(11)[1:]
last_nbs = corrs.loc[uid_last].nlargest(11)[1:]
first_nbs

reviewed_asins_first = df_train[df_train["reviewerID"]==uid_first]["asin"]
reviewed_asins_last = df_train[df_train["reviewerID"]==uid_last]["asin"]
reviewed_asins_first, reviewed_asins_last

#rerun with last_nbs.index to get rate history of the other user's neighbours
nb_reviews = []
for rid in last_nbs.index: 
    nb_reviews.append(df_train[df_train["reviewerID"]==rid][["overall", "reviewerID", "asin", "reviewTime"]])
rating_history = pd.concat(nb_reviews).reset_index(drop=True)
rating_history

df_train[df_train["asin"].isin(reviewed_asins_first)].groupby("asin")["overall"].count()
